Switch from fasd → zoxide
Fasd's last updated was 2015 (which is fine) but the project was archived on GitHub in 2022 and now Homebrew is yelling at me about it. Switching to a more modern version of the same thing written in Rust. Seems like a drop-in replacement that has the same shell aliases for quickly jumping between directories.