Setup a .zshrc file instead of a .zprofile file
I was already using exclusively .zshrc on my Mac and am now bringing that change into the dotfiles project. There's good reasons to put certain things in zprofile (which runs once on user login) and other things in zshrc (which runs on every new terminal session) but using only a zshrc file is more straightforward to manage and it all runs very quickly — certainly makes iterating on dotfiles easier because I just have to open a new Terminal window for all of the changes to be re-applied.