Add convenience script for C development
This is a simple script intended to be used like this:

    $ c hello

Which would them open Vim, so you can edit `hello.c`, and when
you save the file with `:wq`, it will automatically compile the
file, then run it.

The goal is a fast and easy feedback loop for iterating on C files.