Auto switch Node using .nvmrc file
This is the official script recommended by the NVM project’s README. I already have a version of this that uses `.node-version`, but this is for projects that use `.nvmrc` instead. It will run before the look for `.node_version`, which I don’t need anymore for any project I’m currently working on, but am leaving for the time being, in service of having a theoretically more complete dotfiles.